% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_labtaxa}
\alias{cache_labtaxa}
\alias{load_labtaxa}
\title{Cache Object}
\usage{
cache_labtaxa(
  x,
  filename = "cached-LDM-SPC.rds",
  destdir = tools::R_user_dir(package = "labtaxa")
)

load_labtaxa(
  filename = "cached-LDM-SPC.rds",
  destdir = ldm_data_dir(),
  silent = FALSE
)
}
\arguments{
\item{x}{A \code{SoilProfileCollection} (or other object to cache in \code{labtaxa} directory)}

\item{filename}{Default: \code{"cached-LDM-SPC.rds"}}

\item{destdir}{Default: \code{tools::R_user_dir(package = "labtaxa")}}

\item{silent}{Suppress error messages? Passed to \code{try()}. Default \code{FALSE}}
}
\value{
\code{cache_labtaxa()}: \code{x} is cached in \code{destdir} as \code{filename} as a side-effect. Invisibly returns the full file path.

\code{load_labtaxa()}: Read a .rds file (\code{filename}) from \code{destdir} (the labtaxa "user data" directory). Returns the object contained in the file, or \code{try-error} on failure to read file.
}
\description{
Cache Object
}
\examples{
\dontrun{
 x <- get_LDM_snapshot()
 cache_labtaxa(x)
 y <- load_labtaxa()
 all(profile_id(x) == profile_id(y))
}
}
