% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{get_LDM_snapshot}
\alias{get_LDM_snapshot}
\title{Get 'Lab Data Mart' 'SQLite' Snapshot}
\usage{
get_LDM_snapshot(
  ...,
  basename = "NCSSLabDataMartSQLite.sqlite3",
  dirname = tools::R_user_dir(package = "labtaxa"),
  default_dir = "~/Downloads",
  port = 4567L,
  baseurl = "https://ncsslabdatamart.sc.egov.usda.gov/database_download.aspx"
)
}
\arguments{
\item{...}{Arguments passed to \code{soilDB::fetchLDM()}}

\item{basename}{Default: \code{"NCSSLabDataMartSQLite.sqlite3"}}

\item{dirname}{Data cache diretory for \code{labtaxa} package. Default: \code{tools::R_user_dir(package = "labtaxa")}.}

\item{default_dir}{Default directory where RSelenium Gecko (Firefox) downloads files. Default: \code{"~/Downloads"}. Customize as needed.}

\item{port}{Default: \code{4567L}}

\item{baseurl}{Default: \code{"https://ncsslabdatamart.sc.egov.usda.gov/database_download.aspx"}}
}
\value{
A SoilProfileCollection
}
\description{
Downloads the SQLite database snapshot from \code{"https://ncsslabdatamart.sc.egov.usda.gov/database_download.aspx"}. Uses RSelenium to navigate the webpage and download the file, which is cached in a \code{tools::R_user_dir(package="labtaxa")}. Lboratory pedon data are retrieved from the local SQLite file using \code{soilDB::fetchLDM()}.
}
\examples{
\dontrun{
get_LDM_snapshot()
}
}
