% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labtaxa-package.R
\docType{package}
\name{labtaxa-package}
\alias{labtaxa}
\alias{labtaxa-package}
\title{labtaxa: Fetch and Process USDA-NRCS-NCSS Lab Data Mart 'SQLite' Snapshot}
\description{
Supports automatic download of 'Lab Data Mart' (\url{https://ncsslabdatamart.sc.egov.usda.gov/}) database snapshots. Applies patches to the cached database for testing and use in R.
}
\keyword{internal}
